Suppress GNATprove warnings for "statement has no effect"
These warnings are raised for use of generic Boolean formal parameters,
such as Reflect_Output, in if-statements. The warning is correct for
a specific instantiation, but the if-statements do have an effect for
different generic instantiations, depending on the value of the
generic formal parameter. This is the desired behaviour.